+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Ability to normalize DOIs to https://doi.org URI format
+
 ### Fixed
 - Fixed regex so we don't run the invalid multi-value separator fix on
 `dcterms.bibliographicCitation` fields